requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
evaluate==0.4.5
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
ibm-cos-sdk==2.14.2
ibm-cos-sdk-core==2.14.2
ibm-cos-sdk-s3transfer==2.14.2
ibm_watsonx_ai==1.3.30
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
joblib==1.5.1
lomond==0.3.3
multidict==6.6.3
multiprocess==0.70.16
numpy==2.3.1
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
propcache==0.3.2
pyarrow==20.0.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
unitxt==1.26.1
urllib3==2.5.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
asyncpg==0.30.0
sqlglot==27.0.0
psycopg2-binary==2.9.10
loguru==0.7.3
func-timeout==4.3.5
pytest==8.4.1
matplotlib==3.10.5
python-dotenv==1.1.1
pyhive==0.7.0
presto-python-client==0.8.4
langgraph>=0.2.0
langchain-core>=0.3.0
aiomysql>=0.2.0
greenlet>=3.0.0